Move JGitClient to a top level file
commit-id: I303eb3d7
